% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot.R
\name{myplot}
\alias{myplot}
\title{myplot}
\usage{
myplot(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}
}
\value{
a plot
}
\description{
myplot
}
\examples{
# Here is an example
set.seed(312)
x <- rnorm(100)
y <- rnorm(100)
myplot(x, y)

}
