% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ps_matchCPP}
\alias{ps_matchCPP}
\title{Propensity score match from Rcpp}
\usage{
ps_matchCPP(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
A list with two vectors (\code{match_id}, \code{match_x}).
}
\description{
Propensity score match from Rcpp
}
